pyproject.toml
[project]
name = "human-use"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0",
    "rapidata>=2.0",
    "pydantic[email]>=2.0",
    "python-dotenv>=1.0",
    "fastapi>=0.100",
    "sse-starlette>=2.0",
    "anthropic>=0.30",
    "uvicorn>=0.24",
    "httpx>=0.28.1",
    "sqlmodel>=0.0.21",
    "asyncpg>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.13",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
]

[project.scripts]
human-use = "human_use.server:main"

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24", "pandas>=2.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/human_use"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pandas>=2.3.3",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "httpx>=0.27",
    "aiosqlite>=0.20",
]
